requirements.txt
arrow==1.3.0
attrs==23.2.0
blinker==1.7.0
bravado==11.0.3
bravado-core==6.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
fqdn==1.5.1
idna==3.6
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
monotonic==1.6
msgpack==1.0.7
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
simplejson==3.19.2
six==1.16.0
swagger-spec-validator==3.0.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==2.1.0
webcolors==1.13
Werkzeug==3.0.1
